service1/testing/requirements.txt
Flask
pytest
flask_testing
pytest-cov
requests_mock
